Modernisation and code coverage
The project is modernized in patches, almost like the maintainers are
afraid to break anything.  It's using Babel 7, with Stage Zero preset
(meaning *all* of the latest JS features are being transpiled), but not
taking advantage of a lot of those features (cf the `events.jsx` file;
it makes more sense to do its handlers object as a Map/WeakMap than a
raw object, especially if it'll get transpiled down to that anyway).

Other features added include a better `.gitignore` file with Node in
mind, an EditorConfig file so that editing environs can be unified (with
the appropriate plugins, see <https://editorconfig.org> for details),
ESLint for any new mistakes that may pop up, Prettier for unified code
formatting (feel free to modify it in a `.prettierrc` file, btw),
`husky` and `lint-staged` for running Prettier against the modified
files (hopefully all files get touched so the codebase ends up unified
in appearance), Lodash as an awesome utility kit, and Istanbul and NYC
for code coverage.

Code coverage was surprisingly high for a project that wasn't tracking
it before: just on the initial run of `yarn coverage`, it was already
nearly 72 percent.  The file with the lowest coverage was `events.jsx`,
which is what made me open the file in the first place.  Another file to
check is the `NodeHistoryAPI`, which was almost as low.

So, the to-list is as follows (for me, anyway):

[ ] Bring code coverage to over 95%
    [ ] Bring code coverage to over 75%
    [ ] Bring code coverage to over 80%
    [ ] Bring code coverage to over 85%
    [ ] Bring code coverage to over 90%
[ ] Check the code for more modernization opportunities
[ ] Try to enhance the code's FP style
    [ ] Try to reduce any mutation seen (haven't looked through all the
    code yet, though)
[ ] Further modernization
[ ] Aim for 100% code coverage?

Signed-off-by: Jonathan Sifuentes <jayands2k11@ymail.com>