Code coverage (mostly) at 75% or better
I've fully integrated the code coverage suite into the cycle.  Or
rather, I _had_ done so.  I was unable to make a push because of it,
initially; after doing so, I fully intend to reintroduce it in the
cycle.

Husky allows for a "prepush" key in the "scripts" that gets run whenever
you go to `git push`.  Only a clean exit will allow you to successfully
continue.  Since the branch percentage is still in the low 50s, I
couldn't get it to actually let me through.  So the "prepush" script is
actually supposed to read `npm run coverage && lint-staged`.

That said, all the other code coverage is near 80%.

Also, I needed flow for a few files (including some that I then stripped
it out of, most notably the `RouteHistory` that has decorators, because
flow didn't like them), and the `flow-typed/npm` cache in specific is
ignored, so if type declarations are written, they can go in the
top-level `flow-typed` folder and will be used properly.  To get the
type definitions, run the script "types".

I use the lodash plugin for eslint because it knows all the tricks that
I don't; for example `_.compact` instad of filtering on a boolean value.

The recommended extensions are things that I think will enhance the
coding experience in VSCode.  All of them are, of course, optional, but
I would implore other users to at _least_ install the EditorConfig one,
as that's what makes `.editorconfig` files useful, is people actually
enacting the rules imposed.

I'm sure I've forgotten more that I've added/modified (c.f the code
coverage output), but that seems like a decent summary.

Signed-off-by: Jonathan Sifuentes <jayands2k11@ymail.com>