Add Rollup as a build system
This commit adds a `config/rollup` folder, and in it, a common config to
be shared and the two configurations to be ran by rollup: a development
environment one which does not minify the output, and a production one
that does.  I did it that way to avoid duplication, since the majority
of the settings are the same for both.  As a result, chances are the
file you want to look at/edit is the file `rollup.config.common.js` for
the most pertinent settings.

I also realized that I forgot to include a few things in the
`.eslintrc.js` and that I forgot to include my usual `lint` and
`lint:fix` scripts.  This patch also fixes that, so linting can be done
on the command line as well.

Signed-off-by: Jonathan Sifuentes <jayands2k11@ymail.com>