Fix the test build
Rollup's settings of `modules: false, useBuiltins: false` for the env
preset breaks the build for mocha, so I just copied the settings to the
test environment and removed them.  I'm glad I caught it with the
prepush hook, though; it would have been bad to have travis fail like
that.

Signed-off-by: Jonathan Sifuentes <jayands2k11@ymail.com>